reusable-streaming (#5709)
